add boost vector and entity state base class
I'm replacing the custom math library with the boost
equivalents. Additionally I hope to create the find
closest element process by mapping the elements in a normal
map ordered by a hash computed with the cantor pairing function.
While this slows down normal look up times, the majority of the
loop ups in the map are likely to be based upon some arbitrary
point in 3 space relative to the grid's center, so I'm hoping that
the end result is a faster overall lookup. Additionally, I'm
intending to attach state information to tiles with string keys
and realizations of the respective BaseState interfaces.