Rewrite grid classes using vector and shared_ptr
The corner, edge, tile, and grid classes were
refactored as true classes with attached helper functions.
I replaced the rampant raw pointers with shared_ptrs and
weak_ptrs so that I don't have to worry about memory management.
Finally This includes the implementation of the creation of
the base dodecahedron grid.